Two tatches
The first patch makes all the mobs in the world the same level as you (only those that are lower level than you). The mob becomes your level after the first attack on it. The mod works well with the "autobalance" module. Also, all quests become your level and for them the correct amount of experience is given for your level.

The second patch adds random buffs to all things (level "green" and above). At me it is added directly to a thing. These items can be enchanted just like ordinary items. Diablo style.